joblist done

+++ routes/web.php
@@ -20,37 +20,39 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [ListingController::class, 'index']);
 
 //Create Listing
-Route::get('/create', [ListingController::class, 'create']);
+Route::get('/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
-
-// Register form
-Route::get('/register', [UserController::class, 'create']);
-
-
-//Create user
-Route::post('/register', [UserController::class, 'store']);
-
-
-// Signin form
-Route::get('/login', [UserController::class, 'login_form']);
-
-// Login user
-Route::post('/users/authenticate', [UserController::class, 'authenticate']);
-
-// Signout user
-Route::post('/logout', [UserController::class, 'logout']);
-
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'delete']);
+Route::delete('/listings/{listing}', [ListingController::class, 'delete'])->middleware('auth');
+
+//Mange Listing
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Single listing
 Route::get('/listings/{listing}', [ListingController::class, 'listing']);
+
+
+
+//Create user
+Route::post('/register', [UserController::class, 'store'])->middleware('guest');
+
+// Signin form
+Route::get('/login', [UserController::class, 'login_form'])->name('login')->middleware('guest');
+
+// Signout user
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Register form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// Authenticate user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
